Update - em cadastroVIEW para que seja possivel cadastrar produtos atraves do botao 'salvar'

+++ LeiloesTDSat/src/beans/ProdutosDTO.java
@@ -3,7 +3,7 @@ package beans;
 public class ProdutosDTO {
     private Integer id;
     private String nome;
-    private Integer valor;
+    private Double valor;
     private String status;
 
     public Integer getId() {
@@ -14,7 +14,7 @@ public class ProdutosDTO {
         return nome;
     }
 
-    public Integer getValor() {
+    public Double getValor() {
         return valor;
     }
 
@@ -30,7 +30,7 @@ public class ProdutosDTO {
         this.nome = nome;
     }
 
-    public void setValor(Integer valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
